factor out and genericize tezos constants request
