+++ relay/src/implementations/rpc/rpc-cycle-monitor.ts
@@ -1,65 +1,15 @@
-import * as http from "http";
-import { max } from "lodash";
 import CycleMonitor from "../../interfaces/cycle-monitor";
 import GenericCycleMonitor from "../in-memory/generic-cycle-monitor";
 import RpcBlockMonitor from "./rpc-block-monitor";
+import ConstantsUtil from "./rpc-constants";
 
 
 export default class RpcCycleMonitor extends GenericCycleMonitor implements CycleMonitor {
-  private static handleError(rpcApiUrl: string,
-                        retryTimeout: number,
-                        maxRetries: number,
-                        message: string,
-                        resolve: (value: number | PromiseLike<number>) => void,
-                        reject: (reason?: any) => void)
-  {
-    console.error(message);
-    if (maxRetries > 0) {
-      setTimeout(() => {
-        console.error(`Retrying constants request, retries left: ${--maxRetries}`);
-        resolve(RpcCycleMonitor.getBlocksPerCycle(rpcApiUrl, retryTimeout, maxRetries));
-      }, retryTimeout);
-    } else {
-      reject(`Error while fetching or parsing network constants: ${message}`);
-    }
-  }
-
   private static async getBlocksPerCycle(rpcApiUrl: string,
                                             retryTimeout = 1000,
                                             maxRetries = 1000): Promise<number>
   {
-    return new Promise((resolve, reject) => 
-      http.get(`${rpcApiUrl}/chains/main/blocks/head/context/constants`, (resp) => {
-        const { statusCode } = resp;
-        const contentType = resp.headers['content-type'] || '';
-
-        var error;
-        if (statusCode !== 200) {
-          error = new Error(`Constants request failed with status code: ${statusCode}.`);
-        } else if (!/^application\/json/.test(contentType)) {
-          error = new Error(`Constants request produced unexpected response content-type ${contentType}.`);
-        }
-        if (error) {
-          resp.resume();
-          this.handleError(rpcApiUrl, retryTimeout, maxRetries, error.message, resolve, reject);
-          return;
-        }
-
-        // A chunk of data has been received.
-        var rawData = '';
-        resp.on('data', (chunk) => { rawData += chunk; });
-        resp.on('end', () => {
-          try {
-            resolve((JSON.parse(rawData) as {blocks_per_cycle: number}).blocks_per_cycle);
-          } catch (e) {
-            var errMessage = (typeof e === "string") ? e : (e instanceof Error) ? e.message : '';
-            this.handleError(rpcApiUrl, retryTimeout, maxRetries, errMessage, resolve, reject);
-          }
-        });
-      }).on("error", (err) => {
-        this.handleError(rpcApiUrl, retryTimeout, maxRetries, err.message, resolve, reject);
-      })
-    )
+    return ConstantsUtil.getConstant('blocks_per_cycle', rpcApiUrl, retryTimeout, maxRetries);
   }
 
   constructor(
